+++ internal/api/block_documents.go
@@ -2,9 +2,6 @@ package api
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
 
 	"github.com/google/uuid"
 )
@@ -50,60 +47,6 @@ type BlockDocumentUpdate struct {
 	MergeExistingData bool       `json:"merge_existing_data"`
 }
 
-// AccessControlList is a custom type that represents
-// an API response where the value can be:
-// []uuid.UUID - a list of IDs
-// ["*"] - a wildcard, meaning "all".
-//
-// nolint:musttag // we have custom marshal/unmarshal logic for this type
-type AccessControlList struct {
-	IDs []uuid.UUID
-	All bool
-}
-
-// Custom JSON marshaling for AccessControlList
-// so we can return []uuid.UUID or ["*"] back to the API.
-func (acl AccessControlList) MarshalJSON() ([]byte, error) {
-	if acl.All {
-		data, err := json.Marshal([]string{"*"})
-		if err != nil {
-			return nil, fmt.Errorf("failed to marshal wildcard ACL: %w", err)
-		}
-
-		return data, nil
-	}
-
-	data, err := json.Marshal(acl.IDs)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal ID slice ACL: %w", err)
-	}
-
-	return data, nil
-}
-
-// Custom JSON unmarshaling for AccessControlList
-// so we can accept []uuid.UUID or ["*"] from the API
-// in a structured format.
-func (acl *AccessControlList) UnmarshalJSON(data []byte) error {
-	var ids []uuid.UUID
-	if err := json.Unmarshal(data, &ids); err == nil {
-		acl.IDs = ids
-		acl.All = false
-
-		return nil
-	}
-
-	var all []string
-	if err := json.Unmarshal(data, &all); err == nil && len(all) == 1 && all[0] == "*" {
-		acl.All = true
-		acl.IDs = nil
-
-		return nil
-	}
-
-	return errors.New("invalid AccessControlList format")
-}
-
 // BlockDocumentAccessReplace is the "update" request payload
 // to modify a block document's current access control levels,
 // meaning it contains the list of actors/teams + their respective access
@@ -113,24 +56,6 @@ type BlockDocumentAccessReplace struct {
 	ViewActorIDs   AccessControlList `json:"view_actor_ids"`
 	ManageTeamIDs  []uuid.UUID       `json:"manage_team_ids"`
 	ViewTeamIDs    []uuid.UUID       `json:"view_team_ids"`
-}
-
-// BlockActorType represents an enum of type values
-// used in the Block Document Access API.
-type BlockActorType string
-
-const (
-	BlockUser           BlockActorType = "user"
-	BlockServiceAccount BlockActorType = "service_account"
-	BlockTeam           BlockActorType = "team"
-	BlockAll            BlockActorType = "*"
-)
-
-type ObjectActorAccess struct {
-	ID    AccessControlList `json:"id"`
-	Name  string            `json:"name"`
-	Email *string           `json:"email"`
-	Type  BlockActorType    `json:"type"`
 }
 
 // BlockDocumentAccess is the API object representing a
